Revert "Move model into map folder"
This reverts commit fbe1a9b92607b4afc71c64307e2ba93fd9da0f72.
