+++ include/radix/model/MeshLoader.hpp
@@ -4,7 +4,7 @@
 #include <string>
 #include <map>
 
-#include <radix/map/model/Mesh.hpp>
+#include <radix/model/Mesh.hpp>
 
 struct aiMesh;
